webcachesim_server/requirements.txt
asgiref==3.2.3
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.2.1
djangorestframework==3.11.0
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
idna==2.9
importlib-metadata==1.5.0
jsonschema==3.2.0
paramiko==2.7.1
psycopg2==2.8.4
pycparser==2.19
PyNaCl==1.3.0
pyrsistent==0.15.7
pytz==2019.3
PyYAML==5.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
texttable==1.6.2
urllib3==1.25.8
websocket-client==0.57.0
zipp==3.0.0
